+++ src/main/java/com/stoury/exception/member/MemberSearchException.java
@@ -1,0 +1,10 @@
+package com.stoury.exception.member;
+
+public class MemberSearchException extends RuntimeException {
+    public MemberSearchException() {
+    }
+
+    public MemberSearchException(String message) {
+        super(message);
+    }
+}
